requirements.txt
django~=4.1.5
djangorestframework~=3.14.0
psycopg2-binary
django-filter~=22.1

python-dotenv~=0.21.1